+++ HOTEL_USER.py
@@ -1,0 +1,3 @@
+from system_procedures import *
+from system_objects import *
+system_clear_days()
