+++ login_page.py
@@ -1,0 +1,6 @@
+from system_procedures import *
+from system_objects import *
+Calender.load_calender()
+
+#log in
+username = input("username:") #'jonnylil12'
