+++ sign_up_page.py
@@ -1,0 +1,3 @@
+from system_procedures import *
+from system_objects import *
+Calender.load_calender()
